Add an option to merge observed and future sequences for encoding z
This is a needed addition, since in the same-context toy dataset experiment, I am splitting a sequence into observed and future parts with a 50-50 split. But the crucial part of the information is in the last 25% of the data, where either the sinusoid curve flattens or keeps oscilating. Therefore, if only the observed part is used for encoding, then not enough information is present for z to learn anything. Therefore, when encoding the context (and in training - the target as well), we should be encoding also the future part of the sequence.